If listing buckets fails, it maybe due to lack of permissions, so try listing a non-existent bucket instead and see how that fails

+++ save_s3.cgi
@@ -59,8 +59,11 @@ else {
 	# Validate that it works
 	my $buckets = &s3_list_buckets($s3);
 	if (!ref($buckets)) {
-		&delete_s3_account($s3) if ($in{'new'});
-		&error(&text('s3_echeck', $buckets));
+		my $files = &s3_list_files($s3, undef, &random_password(128));
+		if (!ref($files) && $files !~ /does\s+not\s+exist/i) {
+			&delete_s3_account($s3) if ($in{'new'});
+			&error(&text('s3_echeck', $buckets));
+			}
 		}
 
 	# Save the account
